// arcadia.game.core.api/src/modules/messaging/player.handling/enum/player.message.type.ts
export enum PlayerMessageType {
  DEFAULT = 'message',
  PLAYER_JOINED = 'userJoined',
  PLAYER_LEFT = 'userLeft',
  BUY = 'buy',
  QUEUE = 'queue',
  WIN = 'win',
  TOTAL_WIN = 'totalWin',
  SESSION_RESULT = 'sessionResult',
  SESSION_STATE = 'sessionState',
  QUEUE_BALANCE = 'queueBalance',
  RE_BUY = 'reBuy',
  NOTIFICATION = 'notification',
  PLAYER_TO_ROBOT = 'player2robot',
  AUTOPLAY = 'autoplay',
  RETURN_TO_LOBBY = 'returnToLobby',
  PHANTOM = 'phantom',
  BETS = 'bets',
  RESET_IDLE_TIMEOUT = 'resetIdleTimeout',
  BALANCE = 'balance',
  JACKPOT = 'jackpot',
  VOUCHER = 'voucher',
  ROUND_START = 'roundStart',
  REMAINING_COINS = 'remainingCoins',
  FORCE_RECONNECT = 'forceReconnect',
  READY_FOR_NEXT_ROUND = 'readyForNextRound',
  DISABLE_AUTOPLAY = 'disableAutoplay',
  DISABLE_BET_BEHIND = 'disableBetBehind',
}